// ClinicApp.FrontEnd/src/app/pages/billing/models/excel-data.model.ts
export interface ExcelData {
  startDate: string;
  endDate: string;
  primaryPayer: string;
  insurancePlan: string;
  recipientID: string;
  recipientName: string;
  accountNumber: string;
  referringPhysicianNPI: string;
  pos: string;
  emergency: string;
  diagnosisCode: string;
  renderingProvider: string;
  credential: string;
  npi: string;
  procedure: string;
  totalUnits: string;
  billedAmount: string;
  currency: string;
  claimStatus: string;
  claimReference: string;
}